+++ src/types/index.ts
@@ -1,6 +1,7 @@
 export type { IUser, IUserPicture } from './user'
 export { Roles } from './user'
 export type { IClient, TClientName } from './client'
+export type { IAddress, IAddressPreview } from './address'
 export type {
   PaginatedResponse,
   TokenPaginationResponse,